refactor: rename java classes to better represent what they do
