+++ scrapers/jobs.cron
@@ -1,2 +1,2 @@
-*/6 * * * * /usr/local/openjdk-17/bin/java -cp /usr/myjava/scrapers.jar nl.esciencecenter.rsd.scraper.language.Main > /proc/$(cat /var/run/crond.pid)/fd/1 2>&1
-3-59/6 * * * * /usr/local/openjdk-17/bin/java -cp /usr/myjava/scrapers.jar nl.esciencecenter.rsd.scraper.LicensesMain > /proc/$(cat /var/run/crond.pid)/fd/1 2>&1
+*/6 * * * * /usr/local/openjdk-17/bin/java -cp /usr/myjava/scrapers.jar nl.esciencecenter.rsd.scraper.MainProgrammingLanguages > /proc/$(cat /var/run/crond.pid)/fd/1 2>&1
+3-59/6 * * * * /usr/local/openjdk-17/bin/java -cp /usr/myjava/scrapers.jar nl.esciencecenter.rsd.scraper.MainLicenses > /proc/$(cat /var/run/crond.pid)/fd/1 2>&1
