+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/MainLicenses.java
@@ -2,21 +2,15 @@ package nl.esciencecenter.rsd.scraper;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import nl.esciencecenter.rsd.scraper.language.FilterUrlOnlyPLRDecorator;
-import nl.esciencecenter.rsd.scraper.language.GithubPL;
-import nl.esciencecenter.rsd.scraper.language.LicenseData;
-import nl.esciencecenter.rsd.scraper.language.OrderByDatePLRDecorator;
-import nl.esciencecenter.rsd.scraper.language.PostgrestPLR;
-import nl.esciencecenter.rsd.scraper.language.ProgrammingLanguagesRepository;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
 
-public class LicensesMain {
+public class MainLicenses {
 
 	public static void main(String[] args) {
 		System.out.println("Start scraping licenses");
-		ProgrammingLanguagesRepository existingLicensesSorted = new OrderByDatePLRDecorator(new FilterUrlOnlyPLRDecorator(new PostgrestPLR(Config.backendBaseUrl() + "/repository_url?select=id,software,url,license_scraped_at)"), "https://github.com"));
+		SoftwareInfoRepository existingLicensesSorted = new OrderByDateSIRDecorator(new FilterUrlOnlySIRDecorator(new PostgrestSIR(Config.backendBaseUrl() + "/repository_url?select=id,software,url,license_scraped_at)"), "https://github.com"));
 		Collection<LicenseData> dataToScrape = existingLicensesSorted.licenseData();
 		JsonArray allDataToSave = new JsonArray();
 		String scrapedAt = LocalDateTime.now().toString();
@@ -30,7 +24,7 @@ public class LicensesMain {
 				String repo = repoUrl.replace("https://github.com/", "");
 				if (repo.endsWith("/")) repo = repo.substring(0, repo.length() - 1);
 
-				String scrapedJsonData = new GithubPL("https://api.github.com").license(repo);
+				String scrapedJsonData = new GithubSI("https://api.github.com").license(repo);
 				JsonObject newData = new JsonObject();
 				newData.addProperty("id", licenseData.id());
 //				we have to add all existing columns, otherwise PostgREST will not do the UPSERT
@@ -43,7 +37,7 @@ public class LicensesMain {
 				e.printStackTrace();
 			}
 		}
-		new PostgrestPLR(Config.backendBaseUrl() + "/repository_url").save(allDataToSave.toString());
+		new PostgrestSIR(Config.backendBaseUrl() + "/repository_url").save(allDataToSave.toString());
 		System.out.println("Done scraping licenses");
 	}
 }
