+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/RepositoryUrlData.java
@@ -1,4 +1,4 @@
-package nl.esciencecenter.rsd.scraper.language;
+package nl.esciencecenter.rsd.scraper;
 
 import java.time.LocalDateTime;
