+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/SoftwareInfo.java
@@ -1,0 +1,8 @@
+package nl.esciencecenter.rsd.scraper;
+
+public interface SoftwareInfo {
+
+	String data(String repo);
+
+	String license(String repo);
+}
