--- scrapers/src/main/java/nl/esciencecenter/rsd/scraper/language/ProgrammingLanguages.java
@@ -1,8 +1,0 @@
-package nl.esciencecenter.rsd.scraper.language;
-
-public interface ProgrammingLanguages {
-
-	String data(String repo);
-
-	String license(String repo);
-}
